requirements.txt
SQLAlchemy~=2.0.41
dotenv~=0.9.9
python-dotenv~=1.1.0
Flask~=3.1.1
Flask-JWT-Extended~=4.7.1
bcrypt~=4.3.0
# python -c "import secrets; print(secrets.token_urlsafe(64))"
reportlab~=4.4.2
gunicorn~=23.0.0
flask-cors~=6.0.1
pillow_heif~=1.1.0
pillow~=11.2.1
boto3~=1.40.44
botocore~=1.40.44